Advanced Player Done
The advanced movement is now finished for the player and ready for user testing.

Wall jump has been changed to make it more horizontal rather than vertical.

File structure has been cleaned, removing "player" from scripts relating to the player as it is clear for file structure.

New animations for the player need to be added along with the fixing the animation bugs mentioned before.

The shotgun still doesn't fire pellets or render when being used.