Added HUD and scripted boss levels
A HUD has been created to track death, time in the level and if the shotgun can be used.
The ability to skip levels on dying has been added.
The scripts needed for the boss events on bosses 2 and 3 have been created, the levels can now be built.
With this the base gameplay elements are full programmed and ready for further testing.