Player Rewrite and other changes
This contains the player rewrite and changes to player feel.
The bugs relating to the player have all been fixed.
I forgot to GIT after the player rewrite was complete to this also contains the menu, with the level select feature and the latest 3 level and boss needed for the fourth test